requirements.txt
youtube-transcript-api
sentence-transformers
pymilvus
requests
python-dotenv
validators
flask